Test more shenanigans with doubles

+++ matrix.cpp
@@ -5,7 +5,7 @@ matrix::matrix(int rows, int cols) : data(rows, std::vector<double>(cols)) {}
 
 std::ostream& operator<<(std::ostream& out, matrix& mat) {
     for(std::vector<double> row : mat.data) {
-        for(int item : row) {
+        for(double item : row) {
             out << item << " ";
         }
         out << "\n";
